feat: added jikan endpoints to api
