+++ README.md
@@ -1,0 +1,29 @@
+# AniTrack API
+
+Using Java Springboot, this RESTful API makes calls to the 
+open source Jikan API to retrieve information about all things anime/manga!
+This API is used by the AniTrack UI where users can create an account
+and add all their anime to their tracking portfolio!
+
+## AniTrack UI
+If you want to take a look at the AniTrack UI repo, go ahead and follow
+this link: 	[AniTrackUI](https://www.example.com)
+
+# Running AniTrack API locally
+To run the API locally, simply clone the repo running this command in the terminal
+within the desired directory:
+```
+git clone https://github.com/Luqman2003/AniTrack
+```
+If you have maven installed, it would be wise to run `mvn clean install` to ensure all the
+dependencies have been installed. 
+
+## IDE
+### IntelIj
+If you are working in IntelIj, it is quite simple. It should recognize that it is a springboot application
+and have the "run" icon located at the top right. You can click that, and it should start the API for you.
+### VSCode
+VSCode is quite similar. If you go to the AniTrackApplication.java file, there should be a play button
+to the left of the main method. Clicking it will run the application for you.
+
+Creators: **Luqman Shibly**, **Octavio Badillo**
